Split SettingsPage and MediaDetailPanel into focused sub-components
SettingsPage.tsx (1427 -> 105 lines) is now split into independent
section components under components/settings/ (Trakt, OMDB, RPDB,
AI, Addon, Data, Preferences, Push, Profile, API token), each owning
its own state/API calls with no cross-section coupling.

MediaDetailPanel.tsx (1075 -> 434 lines) is split into
components/media-detail/ sub-components for cast, seasons, providers,
ratings, check-in/watch-history actions, and the watch-date/check-in
modals, with the remaining DetailPanel composing them and owning the
shared item-level state (cast/providers/seasons/check-in data).

SetupWizard.tsx now imports TraktSettings/StatusBadge directly from
their new locations instead of re-exporting through SettingsPage.

Behavior is unchanged; this is a pure structural refactor for
maintainability.

Closes #171
